remove debug prints, old volatile flag code, add atomic for usrp_ready
the new usrp_ready bool is now shared via reference, the same way as the
make flag. the atomics have been tested to perform as expected, even
with /O2 enabled. in general, also cleaned up alot of commented code
used in debugging this problem. removed the old usrp_make which was used
internally in the ImUsrpUi class, which is no longer necessary.